List of functions

List of functions

AutomotiveDrivingModels.DIR_RIGHT — Constant.

LaneChangeChoice

A choice of whether to change lanes, and what direction to do it in

source

AutomotiveDrivingModels.AbstractFeature — Type.

Features can be extracted from QueueRecords.

They always return a FeatureValue, which allows the encoding of discrete / continuous / missing values, which can also be forced to a Float64.

source

AutomotiveDrivingModels.CollisionCallback — Type.

CollisionCallback

Terminates the simulation once a collision occurs

source

AutomotiveDrivingModels.CurveProjection — Type.

CurveProjection

The result of a point projected to a Curve

source

AutomotiveDrivingModels.Frenet — Type.

Frenet ______ roadind: road index s: distance along lane t: lane offset, positive is to left. zero point is the centerline of the lane. ϕ: lane relative heading

source

AutomotiveDrivingModels.FrenetRelativePosition — Type.

Project the given point to the same lane as the given RoadIndex.

This will return the projection of the point, along with the Δs along the lane from the RoadIndex.

source

AutomotiveDrivingModels.IntelligentDriverModel — Type.

Commonly referred to as IDM

source

AutomotiveDrivingModels.LaneFollowingAccel — Type.

LaneFollowingAccel

Longitudinal acceleration

source

AutomotiveDrivingModels.LatLonAccel — Type.

LatLonAccel

Acceleration in the frenet frame

source

AutomotiveDrivingModels.MOBIL — Type.

MOBIL

See Treiber & Kesting, 'Modeling Lane-Changing Decisions with MOBIL'

source

AutomotiveDrivingModels.NeighborLongitudinalResult — Type.

get_neighbor_index_fore(scene::Scene, vehicle_index::Int, roadway::Roadway)

Return the index of the vehicle that is in the same lane as scene[vehicle_index] and in front of it with the smallest distance along the lane

The method will search on the current lane first, and if no vehicle is found it
will continue to travel along the lane following next_lane(lane, roadway).
If no vehicle is found within `max_distance_fore,` a value of `nothing` is returned instead.

source

AutomotiveDrivingModels.PedestrianLatLonAccel — Type.

PedestrianLatLonAccel

Pedestrian walking action. Acceleration in the Frenet frame, along with desired lane after crossing the street.

source

AutomotiveDrivingModels.SidewalkPedestrianModel — Type.

SidewalkPedestrianModel

Walks along the sidewalk until approaching the crosswalk. Waits for the cars to pass, then crosses.

source

AutomotiveDrivingModels.VehicleDef — Type.

Vehicle definition which contains a class and a bounding box.

source

AutomotiveDrivingModels.gen_stadium_roadway — Method.

gen_stadium_roadway(nlanes::Int, length::Float64)

Generate a roadway that is a rectangular racetrack with rounded corners. length = length of the x-dim straight section for the innermost (leftmost) lane [m] width = length of the y-dim straight section for the innermost (leftmost) lane [m] radius = turn radius [m]

  ______________________
 /                      \ 
|                        |
|                        |
\______________________/

source

AutomotiveDrivingModels.gen_straight_roadway — Function.

gen_straight_roadway(nlanes::Int, length::Float64)

Generate a roadway with a single straight segment whose rightmost lane center starts at starts at (0,0), and proceeds in the positive x direction.

source

AutomotiveDrivingModels.get_curve_index — Method.

get_curve_index(curve::Curve, s::Float64)

Return the CurveIndex for the closest s-location on the curve

source

AutomotiveDrivingModels.get_curve_index — Method.

get_curve_index(ind::CurveIndex, curve::Curve, Δs::Float64)

Return the CurveIndex at ind's s position + Δs

source

AutomotiveDrivingModels.get_first_collision — Method.

Loops through the scene and finds the first collision between any two vehicles

source

AutomotiveDrivingModels.get_first_collision — Method.

Loops through the scene and finds the first collision between a vehicle and scene[target_index]

source

AutomotiveDrivingModels.get_lerp_time — Method.

get_lerp_time(A::VecE2, B::VecE2, Q::VecE2)

Get lerp time t∈[0,1] such that lerp(A, B) is as close as possible to Q

source

AutomotiveDrivingModels.is_potentially_colliding — Method.

A fast collision check to remove things clearly not colliding

source

AutomotiveDrivingModels.move_along — Function.

move_along(roadind::RoadIndex, road::Roadway, Δs::Float64)

Return the RoadIndex at ind's s position + Δs

source

AutomotiveDrivingModels.propagate — Method.

propagate(veh, action, roadway, Δt)

Take an entity of type {S,D,I} and move it over Δt seconds to produce a new entity based on the action on the given roadway.

source

AutomotiveDrivingModels.read_dxf — Method.

read_dxf(io::IO, ::Type{Roadway})

Return a Roadway generated from a DXF file

Layers with names such as seg001 will contain LWPOLYLINEs.
Each LWPOLYLINE corresponds to a lane centerline, which together
are all neighbored.

source

AutomotiveDrivingModels.simulate! — Method.

Run a simulation and store the resulting scenes in the provided QueueRecord.

Only the ego vehicle is simulated; the other vehicles are as they were in the provided trajdata Other vehicle states will be interpolated

source

AutomotiveDrivingModels.simulate! — Method.

Run nticks of simulation and place all nticks+1 scenes into the QueueRecord

source

Vec.lerp — Method.

quadratic bezier lerp

source

Vec.lerp — Method.

cubic bezier lerp

source

Vec.proj — Method.

Vec.proj(posG::VecSE2{Float64}, curve::Curve)

Return a CurveProjection obtained by projecting posG onto the curve

source

Vec.proj — Method.

proj(posG::VecSE2{Float64}, lane::Lane, roadway::Roadway)

Return the RoadProjection for projecting posG onto the lane. This will automatically project to the next or prev curve as appropriate.

source

Vec.proj — Method.

proj(posG::VecSE2{Float64}, seg::RoadSegment, roadway::Roadway)

Return the RoadProjection for projecting posG onto the segment. Tries all of the lanes and gets the closest one

source

Vec.proj — Method.

proj(posG::VecSE2{Float64}, seg::RoadSegment, roadway::Roadway)

Return the RoadProjection for projecting posG onto the roadway. Tries all of the lanes and gets the closest one

source

AutomotiveDrivingModels.CrossingPhase — Type.

CrossingPhase

Crossing phases for SidewalkPedestrianModel.

For a crossing pedestrian, phases go: -2, -1, 0, 1 For a non-crossing pedestrian that pauses at the crosswalk, phases go: -2, -1, 1 For a non-crossing pedestrian that does not pause at the crosswalk, phases go: -2, 1

Model based on Feliciani et al (2017) - A simulation model for non-signalized pedestrian crosswalks based on evidence from on field observation.

source

AutomotiveDrivingModels.get_lerp_time_unclamped — Method.

get_lerp_time_unclamped(A::VecE2, B::VecE2, Q::VecE2)

Get the interpolation scalar t for the point on the line AB closest to Q This point is P = A + (B-A)*t

source

Base.getindex — Method.

lane[ind::CurveIndex, roadway::Roadway]

Accessor for lanes based on a CurveIndex. Note that we extend the definition of a CurveIndex, previously ind.i ∈ [1, length(curve)-1], to:

ind.i ∈ [0, length(curve)]

where 1 ≤ ind.i ≤ length(curve)-1 is as before, but if the index is on the section between two lanes, we use:

ind.i = length(curve), ind.t ∈ [0,1] for the region between curve[end] → next
ind.i = 0,             ind.t ∈ [0,1] for the region between prev → curve[1]

source