States

States

In this section of the documentation we explain the default vehicle state type provided by AutomotiveDrivingModels as well as the data types used to represent driving scene. Most of the underlying structures are defined in Records.jl. The data structure provided in ADM.jl are concrete instances of parametric types defined in Records. It is possible in principle to define your custom state definition and use the interface defined in ADM.jl.

Agent States

Agents are represented by the entity data type provided by Records.jl. The entity data type has three field: a state, a definition, and an id.

The state of an entity usually describes physical quantity such as position and velocity.

Two states data structure are provided.

1D states and vehicles

AutomotiveDrivingModels.State1D — Type.

State1D

A data type to represent one dimensional states

Fields

- `s::Float64` position 
- `v::Float64` speed [m/s]

source

AutomotiveDrivingModels.Vehicle1D — Type.

Vehicle1D

A specific instance of the Entity type defined in Records.jl to represent vehicles in 1d environments.

source

2D states and vehicles

Here we list useful functions to interact with vehicle states and retrieve interesting information like the position of the front of the vehicle or the lane to which the vehicle belong.

AutomotiveDrivingModels.VehicleState — Type.

VehicleState

A default type to represent an agent physical state (position, velocity). It contains the position in the global frame, Frenet frame and the longitudinal velocity

constructors

VehicleState(posG::VecSE2{Float64}, v::Float64) 
VehicleState(posG::VecSE2{Float64}, roadway::Roadway, v::Float64)
VehicleState(posG::VecSE2{Float64}, lane::Lane, roadway::Roadway, v::Float64)
VehicleState(posF::Frenet, roadway::Roadway, v::Float64)

fields

posG::VecSE2{Float64} global position

posF::Frenet lane relative position

v::Float64 longitudinal velocity

source

Vec.lerp — Method.

Vec.lerp(a::VehicleState, b::VehicleState, t::Float64, roadway::Roadway)

Perform linear interpolation of the two vehicle states. Returns a VehicleState.

source

AutomotiveDrivingModels.get_vel_s — Function.

get_vel_s(s::VehicleState)

returns the longitudinal velocity (along the lane)

source

AutomotiveDrivingModels.get_vel_t — Function.

get_vel_t(s::VehicleState)

returns the lateral velocity (⟂ to lane)

source

AutomotiveDrivingModels.move_along — Method.

move_along(vehstate::VehicleState, roadway::Roadway, Δs::Float64;
ϕ₂::Float64=vehstate.posF.ϕ, t₂::Float64=vehstate.posF.t, v₂::Float64=vehstate.v)

returns a vehicle state after moving vehstate of a length Δs along its lane.

source

AutomotiveDrivingModels.Vehicle — Type.

Vehicle

A specific instance of the Entity type defined in Records to represent Vehicles with state VehicleState , definition VehicleDef and id Int64

source

AutomotiveDrivingModels.get_front — Function.

get_front(veh::Vehicle)

returns the position of the front of the vehicle

source

AutomotiveDrivingModels.get_rear — Function.

get_rear(veh::Vehicle)

returns the position of the rear of the vehicle

source

AutomotiveDrivingModels.get_center — Function.

get_center(veh::Vehicle)

returns the position of the center of the vehicle

source

AutomotiveDrivingModels.get_footpoint — Function.

get_footpoint(veh::Vehicle)

returns the position of the footpoint of the vehicle

source

AutomotiveDrivingModels.get_lane — Function.

get_lane(roadway::Roadway, vehicle::Vehicle)
get_lane(roadway::Roadway, vehicle::VehicleState)

return the lane where vehicle is in.

source

Base.convert — Method.

Base.convert(::Type{Vehicle}, veh::Entity{VehicleState, D, Int64}) where D<:AbstractAgentDefinition

Converts an entity in Vehicle (it is converting the agent definition only)

source

Scenes

A Scene represents a collection of vehicles at a given time.

AutomotiveDrivingModels.Scene — Type.

Scene

A Scene is a specific instance of the Frame type defined in Records. It represents a collection of vehicles at a given time.

Constructors

- `Scene(n::Int=100)`
- `Scene(arr::Vector{Vehicle})`

source

AutomotiveDrivingModels.SceneRecord — Type.

SceneRecord

A SceneRecord is a specific instance of the QueueRecord type defined in Records.jl. It represents a collection of Scenes.

constructor

SceneRecord(capacity::Int, timestep::Float64, frame_capacity::Int=100)

source

AutomotiveDrivingModels.Trajdata — Type.

Trajdata

Trajdata is a specific instance of ListRecord defined in Records.jl. It is a collection of Scenes

source