Feature Extraction [WIP]

AutomotiveDrivingModels.jl implements a feature extraction pipeline to extract information from pre-recorded simulation in a QueueRecord or ListRecord. (see Records.jl for info on the data structure)

PRs for documenting feature extraction are more than welcome

Neighbors Features

Here is a list of function to retrieve information about the neighbors of a given vehicle.

AutomotiveDrivingModels.NeighborLongitudinalResult — Type

NeighborLongitudinalResult

A structure to retrieve information about a neihbor in the longitudinal direction i.e. rear and front neighbors on the same lane. If the neighbor index is equal to nothing it means there is no neighbor.

Fields

ind::Union{Nothing, Int64} index of the neighbor in the scene

Δs::Float64 positive distance along the lane between vehicles positions

AutomotiveDrivingModels.get_neighbor_fore_along_lane — Function

get_neighbor_fore_along_lane(scene::EntityFrame, vehicle_index::Int, roadway::Roadway; max_distance_fore::Float64 = 250.0) 
get_neighbor_fore_along_lane(scene::EntityFrame{S,D,I}, roadway::Roadway, tag_start::LaneTag, s_base::Float64, targetpoint_primary::VehicleTargetPoint, targetpoint_valid::VehicleTargetPoint;
                             max_distance_fore::Float64 = 250.0, index_to_ignore::Int=-1) where {S,D<:AbstractAgentDefinition,I}

                             get_neighbor_fore_along_lane(scene::EntityFrame{S,D,I}, vehicle_index::Int, roadway::Roadway, targetpoint_ego::VehicleTargetPoint, targetpoint_primary::VehicleTargetPoint, # the reference point whose distance we want to minimize
targetpoint_valid::VehicleTargetPoint; # the reference point, which if distance to is positive, we include the vehicle
max_distance_fore::Float64 = 250.0 # max distance to search forward [m]
) where {S,D<:AbstractAgentDefinition,I}

Return the index and the longitudinal distance of the vehicle that is in the same lane as scene[vehicleindex] and in front of it with the smallest distance along the lane. The result is returned as a NeighborLongitudinalResult object. The method will search on the current lane first, and if no vehicle is found it will continue to travel along the lane following nextlane(lane, roadway). If no vehicle is found within max_distance_fore, the index takes a value of nothing.

Notes on the optional arguments:

targetpoint_primary::VehicleTargetPoint the reference point whose distance we want to minimize

targetpoint_valid::VehicleTargetPoint the reference point, which if distance to is positive, we include the vehicle

AutomotiveDrivingModels.get_neighbor_fore_along_left_lane — Function

get_neighbor_fore_along_left_lane(scene::EntityFrame, vehicle_index::Int, roadway::Roadway; max_distance_fore::Float64 = 250.0 )
get_neighbor_fore_along_left_lane(
scene::EntityFrame{S,D,I},
vehicle_index::Int,
roadway::Roadway,
targetpoint_ego::VehicleTargetPoint,
targetpoint_primary::VehicleTargetPoint,
targetpoint_valid::VehicleTargetPoint;
max_distance_fore::Float64 = 250.0
) where {S,D<:AbstractAgentDefinition,I}

Returns the information about the front neighbor of a vehicle on its left lane.

Arguments

scene the scene containing all vehicles information

vehicle_index the index of the vehicle in the scene for which we want to find the neighbor (ego vehicle)

roadway the roadway layout

targetpoint_ego a target point on the ego vehicle that will be used to calculate the distance

targetpoint_primary the reference point whose distance we want to minimize

targetpoint_valid the reference point, which if distance to is positive, we include the vehicle

max_distance_fore max distance to search forward [m]

AutomotiveDrivingModels.get_neighbor_fore_along_right_lane — Function

get_neighbor_fore_along_right_lane(scene::EntityFrame, vehicle_index::Int, roadway::Roadway; max_distance_fore::Float64 = 250.0 )
get_neighbor_fore_along_right_lane(
scene::EntityFrame{S,D,I},
vehicle_index::Int,
roadway::Roadway,
targetpoint_ego::VehicleTargetPoint,
targetpoint_primary::VehicleTargetPoint,
targetpoint_valid::VehicleTargetPoint;
max_distance_fore::Float64 = 250.0
) where {S,D<:AbstractAgentDefinition,I}

Returns the information about the front neighbor of a vehicle on its right lane.

Arguments

scene the scene containing all vehicles information

vehicle_index the index of the vehicle in the scene for which we want to find the neighbor (ego vehicle)

roadway the roadway layout

targetpoint_ego a target point on the ego vehicle that will be used to calculate the distance

targetpoint_primary the reference point whose distance we want to minimize

targetpoint_valid the reference point, which if distance to is positive, we include the vehicle

max_distance_fore max distance to search forward [m]

AutomotiveDrivingModels.get_neighbor_rear_along_lane — Function

get_neighbor_fore_along_rear_lane(scene::EntityFrame, vehicle_index::Int, roadway::Roadway; max_distance_fore::Float64 = 250.0 )
get_neighbor_fore_along_rear_lane(
scene::EntityFrame{S,D,I},
vehicle_index::Int,
roadway::Roadway,
targetpoint_ego::VehicleTargetPoint,
targetpoint_primary::VehicleTargetPoint,
targetpoint_valid::VehicleTargetPoint;
max_distance_fore::Float64 = 250.0
) where {S,D<:AbstractAgentDefinition,I}

Returns the information about the rear neighbor of a vehicle on its own lane.

Arguments

scene the scene containing all vehicles information

vehicle_index the index of the vehicle in the scene for which we want to find the neighbor (ego vehicle)

roadway the roadway layout

targetpoint_ego a target point on the ego vehicle that will be used to calculate the distance

targetpoint_primary the reference point whose distance we want to minimize

targetpoint_valid the reference point, which if distance to is positive, we include the vehicle

max_distance_fore max distance to search forward [m]

Missing docstring.

Missing docstring for get_neighbor_rear_along_left_lane. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_neighbor_rear_along_right_lane. Check Documenter's build log for details.

AutomotiveDrivingModels.FrenetRelativePosition — Type

FrenetRelativePosition

Contains information about the projection of a point on a lane. See get_frenet_relative_position.

Fields

origin::RoadIndex original roadindex used for the projection, contains the target lane ID.

target::RoadIndex roadindex reached after projection

Δs::Float64 longitudinal distance to the original roadindex

t::Float64 lateral distance to the original roadindex in the frame of the target lane

ϕ::Float64 angle with the original roadindex in the frame of the target lane

AutomotiveDrivingModels.get_frenet_relative_position — Function

get_frenet_relative_position(posG::VecSE2{Float64}, roadind::RoadIndex, roadway::Roadway;
    max_distance_fore::Float64 = 250.0, # max distance to search forward [m]
    max_distance_rear::Float64 = 250.0, # max distance to search backward [m]
    improvement_threshold::Float64 = 1e-4,
)

Project the given point to the same lane as the given RoadIndex. This will return the projection of the point, along with the Δs along the lane from the RoadIndex. The returned type is a FrenetRelativePosition object.

Lane Features

AutomotiveDrivingModels.get_lane_width — Function

get_lane_width(veh::Entity, roadway::Roadway)

Returns the width of the lane where veh is.

AutomotiveDrivingModels.get_markerdist_left — Function

get_markerdist_left(veh::Entity, roadway::Roadway)

distance of veh to the left marker of the lane

AutomotiveDrivingModels.get_markerdist_right — Function

get_markerdist_left(veh::Entity, roadway::Roadway)

distance of veh to the right marker of the lane