Feature Extraction

AutomotiveDrivingModels.jl implements a feature extraction pipeline to extract information from pre-recorded simulation in a QueueRecord or ListRecord. (see Records.jl for info on the data structure)

Warning

Work in progress, the feature extraction functionalities are being redesigned Follow #72 for more information PRs for documenting feature extraction are more than welcome

Finding neighbors

Finding neighbors of an entity is a common query and can be done using the find_neighbor function. This function allows to find the neighbor of an entity within a scene using a forward search method. The search algorithm moves longitudinally along a given lane and its successors (according to the given road topology). Once it finds a car it stops searching and return the results as a NeighborLongitudinalResult

AutomotiveDrivingModels.NeighborLongitudinalResult — Type

NeighborLongitudinalResult

A structure to retrieve information about a neihbor in the longitudinal direction i.e. rear and front neighbors on the same lane. If the neighbor index is equal to nothing it means there is no neighbor.

Fields

ind::Union{Nothing, Int64} index of the neighbor in the scene

Δs::Float64 positive distance along the lane between vehicles positions

The find_neighbor function accepts different keyword argument to search front or rear neighbors, or neighbors on different lanes.

AutomotiveDrivingModels.find_neighbor — Function

find_neighbor(scene::Frame, roadway::Roawday, ego::Entity; kwargs...)

Search through lanes and road segments to find a neighbor of ego in the scene.

Arguments

scene::Frame the scene containing all the entities

roadway::Roadway the road topology on which entities are driving

ego::Entity the entity that we want to compute the neighbor of.

Keyword arguments

rear::Bool = false set to true to search for rear neighbor, search forward neighbor by default

max_distance::Float64 = 250.0 stop searching after this distance is reached, if the neighbor is further than max_distance, returns nothing

targetpoint_ego::VehicleTargetPoint the point on the ego vehicle used for distance calculation, see VehicleTargetPoint for more info

targetpoint_neighbor::VehicleTargetPoint the point on the neighbor vehicle used for distance calculation, see VehicleTargetPoint for more info

ids_to_ignore::Union{Nothing, Set{I}} = nothing a list of entity ids to ignore for the search, ego is always ignored.

When computing the distance to a neighbor, one might want to choose different reference points on the vehicle (center to center, bumper to bumper, etc...). AutomotiveDrivingModels provides the VehicleTargetPoint types to do so. One can choose among three possible instances to use the front, center, or rear point of a vehicle to compute the distance to the neighbor.

Missing docstring.

Missing docstring for VehicleTargetPoint, VehicleTargetPointCenter, VehicleTargetPointFront, VehicleTargetPointRear. Check Documenter's build log for details.

To get the relative position of two vehicles in the Frenet frame, the get_frenet_relative_position can be used. It stores the result in a FrenetRelativePosition object.

AutomotiveDrivingModels.get_frenet_relative_position — Function

get_frenet_relative_position(veh_fore::Entity, veh_rear::Entity, roadway::Roadway)

return the Frenet relative position between the two vehicles. It projects the position of the first vehicle onto the lane of the second vehicle. The result is stored as a FrenetRelativePosition.

Lower level: getfrenetrelativeposition(posG::VecSE2{Float64}, roadind::RoadIndex, roadway::Roadway; maxdistancefore::Float64 = 250.0, # max distance to search forward [m] maxdistancerear::Float64 = 250.0, # max distance to search backward [m] improvementthreshold::Float64 = 1e-4, )

Project the given point to the same lane as the given RoadIndex. This will return the projection of the point, along with the Δs along the lane from the RoadIndex. The returned type is a FrenetRelativePosition object.

AutomotiveDrivingModels.FrenetRelativePosition — Type

FrenetRelativePosition

Contains information about the projection of a point on a lane. See get_frenet_relative_position.

Fields

origin::RoadIndex original roadindex used for the projection, contains the target lane ID.

target::RoadIndex roadindex reached after projection

Δs::Float64 longitudinal distance to the original roadindex

t::Float64 lateral distance to the original roadindex in the frame of the target lane

ϕ::Float64 angle with the original roadindex in the frame of the target lane

Lane Features

AutomotiveDrivingModels.get_lane_width — Function

get_lane_width(veh::Entity, roadway::Roadway)

Returns the width of the lane where veh is.

AutomotiveDrivingModels.get_markerdist_left — Function

get_markerdist_left(veh::Entity, roadway::Roadway)

distance of veh to the left marker of the lane

AutomotiveDrivingModels.get_markerdist_right — Function

get_markerdist_left(veh::Entity, roadway::Roadway)

distance of veh to the right marker of the lane