Feature Extraction [WIP]

AutomotiveDrivingModels.jl implements a feature extraction pipeline to extract information from pre-recorded simulation in a QueueRecord or ListRecord. (see Records.jl for info on the data structure)

PRs for documenting feature extraction are more than welcome

Neighbors Features

Here is a list of function to retrieve information about the neighbors of a given vehicle.

AutomotiveDrivingModels.NeighborLongitudinalResult — Type

NeighborLongitudinalResult

A structure to retrieve information about a neihbor in the longitudinal direction i.e. rear and front neighbors on the same lane. If the neighbor index is equal to nothing it means there is no neighbor.

Fields

ind::Union{Nothing, Int64} index of the neighbor in the scene

Δs::Float64 positive distance along the lane between vehicles positions

AutomotiveDrivingModels.get_neighbor_fore_along_lane — Function

get_neighbor_fore_along_lane(scene::EntityFrame{S,D,I}, roadway::Roadway, tag_start::LaneTag, s_base::Float64, targetpoint_primary::VehicleTargetPoint, targetpoint_valid::VehicleTargetPoint;
                             max_distance_fore::Float64 = 250.0, index_to_ignore::Int=-1) where {S,D<:AbstractAgentDefinition,I}

Return the index and the longitudinal distance of the vehicle that is in the same lane as scene[vehicleindex] and in front of it with the smallest distance along the lane. The result is returned as a NeighborLongitudinalResult object. The method will search on the current lane first, and if no vehicle is found it will continue to travel along the lane following nextlane(lane, roadway). If no vehicle is found within max_distance_fore, the index takes a value of nothing.

Notes on the optional arguments:

targetpoint_primary::VehicleTargetPoint the reference point whose distance we want to minimize

targetpoint_valid::VehicleTargetPoint the reference point, which if distance to is positive, we include the vehicle

Missing docstring.

Missing docstring for get_neighbor_fore_along_left_lane. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_neighbor_fore_along_right_lane. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_neighbor_rear_along_lane. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_neighbor_rear_along_left_lane. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_neighbor_rear_along_right_lane. Check Documenter's build log for details.

AutomotiveDrivingModels.FrenetRelativePosition — Type

Project the given point to the same lane as the given RoadIndex.

This will return the projection of the point, along with the Δs along the lane from the RoadIndex.

Missing docstring.

Missing docstring for get_frenet_relative_position. Check Documenter's build log for details.

Lane Features

AutomotiveDrivingModels.get_lane_width — Function

get_lane_width(veh::Entity, roadway::Roadway)

Returns the width of the lane where veh is.

AutomotiveDrivingModels.get_markerdist_left — Function

get_markerdist_left(veh::Entity, roadway::Roadway)

distance of veh to the left marker of the lane

AutomotiveDrivingModels.get_markerdist_right — Function

get_markerdist_left(veh::Entity, roadway::Roadway)

distance of veh to the right marker of the lane